Implement trades_history for Kraken and improve the Trade Class
In this commit I implemented the trades_history
for the kraken exchange. A test for the method
was also added. An important change made during
this commit is the addition of the parameter
time to the class Trader. This not only allows
to simplify the implementation of trades_history
but also helps to better characterize a trade.