Implement order_info for the kraken exchange
This commit implements the order_info method for the Kraken
exchange. Because this is the first method that requires access
to private user data, a helper method _private_query was also
added. In addition, facilities to store and retrieve the
API key of the user were added.

The order_info method returns an instance of OrderInformation
which is a placeholder. Full order information will be added in
the future.