Add Assets and Basic Orders
In this commit I added the Class Assets and added as predefined assets
the ones that I use the most. I also made the Class Order, and create
the classes for the following types of orders: Market (simply order);
Limit and TrailingStop. While I briefly tested the classes, I didn't
add such tests in this commit. In the following commits I'll create
more and add them. The three orders are heavily commented and have
examples to help understand them.